雅琳-member相關設定(測試自動生成資料庫表)，id改long

+++ src/main/java/foodelicious/mainpage/MainUpdateController.java
@@ -49,7 +49,7 @@ public class MainUpdateController {
 		if (resultStatus) {
 			m.addAttribute("account", account);
 			m.addAttribute("pwd", pwd);
-			int id = totalDaoService.findId(new Account(account, pwd));
+			Long id = totalDaoService.findId(new Account(account, pwd));
 			session.setAttribute("account", account);
 			session.setAttribute("pwd", pwd);
 			session.setAttribute("userID", id);
